+++ resources/views/errors/403.blade.php
@@ -1,0 +1,5 @@
+@include('errors._seite', [
+    'code' => 403,
+    'titel' => __('Kein Zugriff'),
+    'text' => __('Für diese Seite fehlt dir die Berechtigung.'),
+])