+++ resources/views/errors/404.blade.php
@@ -1,0 +1,5 @@
+@include('errors._seite', [
+    'code' => 404,
+    'titel' => __('Seite nicht gefunden'),
+    'text' => __('Die Adresse gibt es nicht (mehr). Vielleicht wurde der Eintrag gelöscht oder der Link ist veraltet.'),
+])