+++ resources/views/errors/419.blade.php
@@ -1,0 +1,5 @@
+@include('errors._seite', [
+    'code' => 419,
+    'titel' => __('Sitzung abgelaufen'),
+    'text' => __('Die Seite lag zu lange offen. Melde dich erneut an, dann geht es weiter.'),
+])