+++ resources/views/errors/503.blade.php
@@ -1,0 +1,5 @@
+@include('errors._seite', [
+    'code' => 503,
+    'titel' => __('Kurz nicht erreichbar'),
+    'text' => __('Die Anwendung wird gerade gewartet und ist gleich wieder da.'),
+])